+++ scripts/plot_from_file.py
@@ -79,7 +79,7 @@ with open(filename, "rb") as file_handler:
     results = pickle.load(file_handler)
 
 # Plot table
-writer = pytablewriter.MarkdownTableWriter()
+writer = pytablewriter.MarkdownTableWriter(max_precision=3)
 writer.table_name = "results_table"
 writer.headers = results["results_table"]["headers"]
 writer.value_matrix = results["results_table"]["value_matrix"]
@@ -193,7 +193,7 @@ for key in keys:
             warnings.warn(f"{env} not found for normalizing scores, you should update `env_key_to_env_id`")
 
     # Truncate to convert to matrix
-    min_runs = min([len(algo_score) for algo_score in algo_scores])
+    min_runs = min(len(algo_score) for algo_score in algo_scores)
     if min_runs > 0:
         algo_scores = [algo_score[:min_runs] for algo_score in algo_scores]
         # shape: (n_envs, n_runs) -> (n_runs, n_envs)